+++ test/duckdb_test/appender_test.rb
@@ -996,5 +996,26 @@ module DuckDBTest
 
       assert_equal([1, 'foo'], r.first)
     end
+
+    def test_appender_clear
+      skip 'DuckDB::Appender#clear is not available' unless DuckDB::Appender.method_defined?(:clear)
+
+      @con.query('CREATE TABLE t (id INTEGER, name VARCHAR)')
+      appender = @con.appender('t')
+
+      data = [[1, 'John'], [-2, 'Bob'], [3, 'Jane']]
+      data.each do |(id, name)|
+        appender.append(id).append(name).end_row
+        if id.negative?
+          appender.clear
+          next
+        end
+        appender.flush
+      end
+
+      r = @con.query('SELECT * FROM t ORDER BY id')
+
+      assert_equal([[1, 'John'], [3, 'Jane']], r.each.to_a)
+    end
   end
 end